fix load db and home

+++ app/views.py
@@ -1,34 +1,30 @@
 from django.shortcuts import render
-from django.http import HttpResponse, Http404
+from django.http import Http404
 
 import requests
-import json
 
 from .models import Repo
 
-# todo separar o request do home para nao demorar o loading da pagina.
-def home(request):
-    all_repos = {}
-    # if request.GET:
-    #     language = request.GET.get('nome')
-    Repo.objects.all().delete()
-    for language in ('java', 'python', 'ruby', 'rust', 'go'):
-        url = 'https://api.github.com/search/repositories?q=language:'+language+'&sort=stars&order=desc'
-        response = requests.get(url)
-        data = response.json()
-        repos = data['items']
-        for i in repos:
-            repo_data = Repo(
-                name=i['name'],
-                language=i['language'],
-                stars=i['stargazers_count'],
-                forks=i['forks_count'],
-                description=i['description'],
-                url=i['html_url'],
-            )
-            repo_data.save()
-            all_repos = Repo.objects.all().order_by('-stars')
 
+def home(request):
+    if request.GET and 'Atualizar' in request.GET.get('enviar'):
+        Repo.objects.all().delete()
+        for language in ('java', 'python', 'ruby', 'rust', 'go'):
+            url = 'https://api.github.com/search/repositories?q=language:'+language+'&sort=stars&order=desc'
+            response = requests.get(url)
+            data = response.json()
+            repos = data['items']
+            for i in repos:
+                repo_data = Repo(
+                    name=i['name'],
+                    language=i['language'],
+                    stars=i['stargazers_count'],
+                    forks=i['forks_count'],
+                    description=i['description'],
+                    url=i['html_url'],
+                )
+                repo_data.save()
+    all_repos = Repo.objects.all().order_by('-stars')
     return render(request, 'home.html', {
         "repos": all_repos,
     })
